+++ src/View/Polygon.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #include <SFML/Graphics.hpp>
-#include "Point.h"
+#include "../Model/Point.h"
 
 class Polygon : public sf::Shape {
 public:
 	Polygon();
-	virtual unsigned long getPointCount() const;
+	virtual size_t getPointCount() const;
 	virtual sf::Vector2f getPoint(std::size_t index) const;
 	void setPoint(int index, Point<double> p);
